bit_manipulation

+++ bit_manipulation/2-get_bit.c
@@ -7,7 +7,7 @@
  */
 int get_bit(unsigned long int n, unsigned int index)
 {
-	unsigned long int mask = 1UL << index;
+	int mask = 1UL << index;
 
 	if ((n & mask) != 0)
 	{
